refactor: Move ontology request types to slice layer
Move ontology request classes from `messages.v2.responder.ontologymessages`
to `slice.api.v2.ontologies` package to consolidate API request types in
the slice layer. This includes CreateOntologyRequestV2, CreatePropertyRequestV2,
CanDeleteCardinalitiesFromClassRequestV2, DeleteCardinalitiesFromClassRequestV2,
ChangePropertyGuiElementRequest, ChangeClassLabelsOrCommentsRequestV2, and
LabelOrComment enum. Update all imports across the codebase accordingly.
