+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/ontologymessages/OntologyMessagesV2.scala
@@ -15,7 +15,6 @@ import org.eclipse.rdf4j.model.vocabulary.XSD
 import zio.prelude.Validation
 
 import java.time.Instant
-import java.util.UUID
 
 import dsp.constants.SalsahGui
 import dsp.errors.AssertionException
@@ -23,16 +22,13 @@ import dsp.errors.BadRequestException
 import dsp.errors.DataConversionException
 import dsp.errors.InconsistentRepositoryDataException
 import dsp.valueobjects.Iri
-import dsp.valueobjects.Schema
 import org.knora.webapi.*
 import org.knora.webapi.config.AppConfig
-import org.knora.webapi.core.RelayedMessage
 import org.knora.webapi.messages.IriConversions.*
 import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.OntologyConstants.KnoraApiV2Complex
 import org.knora.webapi.messages.OntologyConstants.KnoraApiV2Simple
 import org.knora.webapi.messages.OntologyConstants.Rdfs
-import org.knora.webapi.messages.ResponderRequest.KnoraRequestV2
 import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.store.triplestoremessages.*
@@ -41,137 +37,12 @@ import org.knora.webapi.messages.v2.responder.*
 import org.knora.webapi.messages.v2.responder.ontologymessages.OwlCardinality.KnoraCardinalityInfo
 import org.knora.webapi.messages.v2.responder.standoffmessages.StandoffDataTypeClasses
 import org.knora.webapi.slice.admin.domain.model.KnoraProject.ProjectIri
-import org.knora.webapi.slice.admin.domain.model.User
-import org.knora.webapi.slice.common.KnoraIris
 import org.knora.webapi.slice.common.KnoraIris.OntologyIri
 import org.knora.webapi.slice.common.KnoraIris.PropertyIri
 import org.knora.webapi.slice.common.KnoraIris.ResourceClassIri
 import org.knora.webapi.slice.common.domain.LanguageCode
 import org.knora.webapi.slice.common.domain.LanguageCode.EN
 import org.knora.webapi.slice.ontology.domain.model.Cardinality
-
-/**
- * An abstract trait for messages that can be sent to `ResourcesResponderV2`.
- */
-sealed trait OntologiesResponderRequestV2 extends KnoraRequestV2 with RelayedMessage
-
-/**
- * Requests the creation of an empty ontology. A successful response will be a [[ReadOntologyV2]].
- *
- * @param ontologyName   the name of the ontology to be created.
- * @param projectIri     the IRI of the project that the ontology will belong to.
- * @param isShared       the flag that shows if an ontology is a shared one.
- * @param label          the label of the ontology.
- * @param comment        the optional comment that described the ontology to be created.
- * @param apiRequestID   the ID of the API request.
- * @param requestingUser the user making the request.
- */
-case class CreateOntologyRequestV2(
-  ontologyName: String,
-  projectIri: ProjectIri,
-  isShared: Boolean,
-  label: String,
-  comment: Option[NonEmptyString],
-  apiRequestID: UUID,
-  requestingUser: User,
-)
-
-/**
- * Requests the addition of a property to an ontology. A successful response will be a [[ReadOntologyV2]].
- *
- * @param propertyInfoContent  an [[PropertyInfoContentV2]] containing the property definition.
- * @param lastModificationDate the ontology's last modification date.
- * @param apiRequestID         the ID of the API request.
- * @param requestingUser       the user making the request.
- */
-case class CreatePropertyRequestV2(
-  propertyInfoContent: PropertyInfoContentV2,
-  lastModificationDate: Instant,
-  apiRequestID: UUID,
-  requestingUser: User,
-) extends OntologiesResponderRequestV2
-
-/**
- * FIXME(DSP-1856): Can only remove one single cardinality at a time.
- * Requests a check if the user can remove class's cardinalities. A successful response will be a [[CanDoResponseV2]].
- *
- * @param classInfoContent     a [[ClassInfoContentV2]] containing the cardinalities to be removed.
- * @param lastModificationDate the ontology's last modification date.
- * @param apiRequestID         the ID of the API request.
- * @param requestingUser       the user making the request.
- */
-final case class CanDeleteCardinalitiesFromClassRequestV2(
-  classInfoContent: ClassInfoContentV2,
-  lastModificationDate: Instant,
-  apiRequestID: UUID,
-  requestingUser: User,
-) extends OntologiesResponderRequestV2
-
-/**
- * FIXME(DSP-1856): Can only remove one single cardinality at a time.
- * Requests the removal of a class's cardinalities. A successful response will be a [[ReadOntologyV2]].
- *
- * @param classInfoContent     a [[ClassInfoContentV2]] containing the cardinalities to be removed.
- * @param lastModificationDate the ontology's last modification date.
- * @param apiRequestID         the ID of the API request.
- * @param requestingUser       the user making the request.
- */
-final case class DeleteCardinalitiesFromClassRequestV2(
-  classInfoContent: ClassInfoContentV2,
-  lastModificationDate: Instant,
-  apiRequestID: UUID,
-  requestingUser: User,
-) extends OntologiesResponderRequestV2
-
-/**
- * Requests that the `salsah-gui:guiElement` and `salsah-gui:guiAttribute` of a property are changed.
- *
- * @param propertyIri          the IRI of the property to be changed.
- * @param newGuiObject         the GUI object with the new GUI element and/or GUI attributes.
- * @param lastModificationDate the ontology's last modification date.
- * @param apiRequestID         the ID of the API request.
- * @param requestingUser       the user making the request.
- */
-case class ChangePropertyGuiElementRequest(
-  propertyIri: KnoraIris.PropertyIri,
-  newGuiObject: Schema.GuiObject,
-  lastModificationDate: Instant,
-  apiRequestID: UUID,
-  requestingUser: User,
-) extends OntologiesResponderRequestV2
-
-/**
- * Requests that a class's labels or comments are changed. A successful response will be a [[ReadOntologyV2]].
- *
- * @param classIri             the IRI of the property.
- * @param predicateToUpdate    `rdfs:label` or `rdfs:comment`.
- * @param newObjects           the class's new labels or comments.
- * @param lastModificationDate the ontology's last modification date.
- * @param apiRequestID         the ID of the API request.
- * @param requestingUser       the user making the request.
- */
-case class ChangeClassLabelsOrCommentsRequestV2(
-  classIri: ResourceClassIri,
-  predicateToUpdate: LabelOrComment,
-  newObjects: Seq[LanguageTaggedStringLiteralV2],
-  lastModificationDate: Instant,
-  apiRequestID: UUID,
-  requestingUser: User,
-) extends OntologiesResponderRequestV2
-
-enum LabelOrComment {
-  case Label
-  case Comment
-
-  override def toString: String = this match {
-    case Label   => RDFS.LABEL.toString
-    case Comment => RDFS.COMMENT.toString
-  }
-}
-object LabelOrComment {
-  def fromString(str: String): Option[LabelOrComment] =
-    LabelOrComment.values.find(_.toString == str)
-}
 
 /**
  * Represents assertions about one or more ontology entities (resource classes and/or properties).
