+++ webapi/src/main/scala/org/knora/webapi/responders/v2/OntologyResponderV2.scala
@@ -30,9 +30,15 @@ import org.knora.webapi.slice.admin.domain.model.KnoraProject.ProjectIri
 import org.knora.webapi.slice.admin.domain.model.User
 import org.knora.webapi.slice.admin.domain.service.KnoraProjectService
 import org.knora.webapi.slice.api.v2.ontologies.AddCardinalitiesToClassRequestV2
+import org.knora.webapi.slice.api.v2.ontologies.CanDeleteCardinalitiesFromClassRequestV2
+import org.knora.webapi.slice.api.v2.ontologies.ChangeClassLabelsOrCommentsRequestV2
 import org.knora.webapi.slice.api.v2.ontologies.ChangeGuiOrderRequestV2
+import org.knora.webapi.slice.api.v2.ontologies.ChangePropertyGuiElementRequest
 import org.knora.webapi.slice.api.v2.ontologies.ChangePropertyLabelsOrCommentsRequestV2
 import org.knora.webapi.slice.api.v2.ontologies.CreateClassRequestV2
+import org.knora.webapi.slice.api.v2.ontologies.CreateOntologyRequestV2
+import org.knora.webapi.slice.api.v2.ontologies.CreatePropertyRequestV2
+import org.knora.webapi.slice.api.v2.ontologies.DeleteCardinalitiesFromClassRequestV2
 import org.knora.webapi.slice.api.v2.ontologies.LastModificationDate
 import org.knora.webapi.slice.api.v2.ontologies.ReplaceClassCardinalitiesRequestV2
 import org.knora.webapi.slice.common.KnoraIris.OntologyIri
